+++ src/main/kotlin/com/viartemev/ktor/flyway/FlywayPlugin.kt
@@ -10,7 +10,7 @@ import javax.sql.DataSource
 
 private val flyWayFeatureLogger: Logger = LoggerFactory.getLogger("com.viartemev.ktor.flyway.FlywayFeature")
 
-class FlywayFeature(configuration: Configuration) {
+class FlywayPlugin(configuration: Configuration) {
     private val dataSource = configuration.dataSource
     private val locations = configuration.locations
     private val schemas = configuration.schemas
@@ -26,12 +26,12 @@ class FlywayFeature(configuration: Configuration) {
         }
     }
 
-    companion object Feature : ApplicationFeature<Application, Configuration, FlywayFeature> {
-        override val key = AttributeKey<FlywayFeature>("FlywayFeature")
+    companion object Static : Plugin<Application, Configuration, FlywayPlugin> {
+        override val key = AttributeKey<FlywayPlugin>("FlywayPlugin")
 
-        override fun install(pipeline: Application, configure: Configuration.() -> Unit): FlywayFeature {
+        override fun install(pipeline: Application, configure: Configuration.() -> Unit): FlywayPlugin {
             val configuration = Configuration().apply(configure)
-            val flywayFeature = FlywayFeature(configuration)
+            val flywayFeature = FlywayPlugin(configuration)
             if (configuration.dataSource == null) throw ApplicationConfigurationException("DataSource is not configured")
 
             flyWayFeatureLogger.info("Flyway migration has started")
